feat: add navigation stack
